+++ server_reset_password.go
@@ -33,10 +33,11 @@ func runServerResetPassword(cli *CLI, cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return err
 	}
-	errCh, _ := waitAction(cli.Context, cli.Client(), result.Action)
-	if err := <-errCh; err != nil {
+
+	if err := cli.ActionProgress(cli.Context, result.Action); err != nil {
 		return err
 	}
+
 	fmt.Printf("Password of server %s reset to: %s\n", server.Name, result.RootPassword)
 	return nil
 }